+++ ManageProductsViewController.swift
@@ -231,7 +231,7 @@ class ManageProductsViewController: UIViewController, UITableViewDataSource, UIT
     func onAddGroup(_ group: ProductGroup, onFinish: VoidFunction?) {
     }
     
-    func onAddProduct(_ product: QuantifiableProduct, quantity: Int) {
+    func onAddProduct(_ product: QuantifiableProduct, quantity: Float) {
     }
     
     func onAddItem(_ item: Item) {
